requirements.txt
streamlit==1.55.0
google-genai
beautifulsoup4
playwright
python-docx
reportlab
requests
python-dotenv
pandas
supabase
supabase
supabase
supabase
